refactor: moved all dependency injection modules into bootstrap.py
