+++ src/dioptra/restapi/user/__init__.py
@@ -16,13 +16,10 @@
 # https://creativecommons.org/licenses/by/4.0/legalcode
 """The experiment endpoint subpackage."""
 
-from .dependencies import bind_dependencies, register_providers
 from .errors import register_error_handlers
 from .routes import register_routes
 
 __all__ = [
-    "bind_dependencies",
     "register_error_handlers",
-    "register_providers",
     "register_routes",
 ]
